refactor message store module
